+++ pickLayer/core/identifygeometry.py
@@ -24,11 +24,11 @@ from qgis.core import QgsFeature, QgsVectorLayer
 from qgis.gui import QgsMapCanvas, QgsMapToolIdentify
 from qgis.PyQt.QtCore import pyqtSignal
 from qgis.PyQt.QtGui import QCursor
+from qgis_plugin_tools.tools.i18n import tr
+from qgis_plugin_tools.tools.messages import MsgBar
+from qgis_plugin_tools.tools.resources import plugin_name
 
 from pickLayer.definitions.settings import Settings
-from pickLayer.qgis_plugin_tools.tools.i18n import tr
-from pickLayer.qgis_plugin_tools.tools.messages import MsgBar
-from pickLayer.qgis_plugin_tools.tools.resources import plugin_name
 
 LOGGER = logging.getLogger(plugin_name())